Umbra: add robustness when looking for METADATA.json

+++ src/main/java/org/wikimedia/commons/donvip/spacemedia/service/orgs/AbstractOrgS3Service.java
@@ -93,29 +93,38 @@ public abstract class AbstractOrgS3Service extends AbstractOrgService<S3Media> {
         int count = 0;
         LocalDateTime start = LocalDateTime.now();
         LocalDate doNotFetchEarlierThan = getRuntimeData().getDoNotFetchEarlierThan();
-        List<S3Media> files = s3.getFiles(region, bucket, summary -> toS3Media(bucket, summary),
+        List<S3Media> files = s3.getFiles(region, bucket, null, S3Service.MEDIA_EXT,
+                summary -> toS3Media(bucket, summary),
                 file -> doNotFetchEarlierThan == null || file.getPublicationDate().isAfter(doNotFetchEarlierThan),
                 Comparator.comparing(S3Media::getPublicationDateTime));
         LOGGER.info("Found {} files in {}", files.size(), Duration.between(start, LocalDateTime.now()));
 
         for (S3Media media : files) {
-            try {
-                Pair<S3Media, Integer> result = processS3Media(media, this::enrichS3Media);
-                if (result.getValue() > 0) {
-                    uploadedMedia.add(result.getKey());
+            if (!skipMedia(media, files)) {
+                try {
+                    Pair<S3Media, Integer> result = processS3Media(media, this::enrichS3Media);
+                    if (result.getValue() > 0) {
+                        uploadedMedia.add(result.getKey());
+                    }
+                    ongoingUpdateMedia(start, bucket, count++);
+                } catch (AmazonS3Exception e) {
+                    LOGGER.warn(e.getMessage());
+                    if (e.getStatusCode() == 404) {
+                        repository.delete(media);
+                    }
+                } catch (UploadException | IOException | RuntimeException e) {
+                    LOGGER.error(e.getMessage(), e);
                 }
-                ongoingUpdateMedia(start, bucket, count++);
-            } catch (AmazonS3Exception e) {
-                LOGGER.warn(e.getMessage());
-                if (e.getStatusCode() == 404) {
-                    repository.delete(media);
-                }
-            } catch (UploadException | IOException | RuntimeException e) {
-                LOGGER.error(e.getMessage(), e);
+            } else {
+                LOGGER.info("Media skipped: {}", media);
             }
         }
 
         return Pair.of(count, uploadedMedia);
+    }
+
+    protected boolean skipMedia(S3Media media, List<S3Media> files) {
+        return false;
     }
 
     protected S3Media toS3Media(String bucket, S3ObjectSummary summary) {
